requirements.txt
# Required dependencies and Modules
customtkinter
scikit-learn
opencv-python
capture_devices
pillow
numpy
